Refactoriza exportador PDF en módulos internos
