+++ viewer2d/pdf/layout_pdf_exporter.cpp
@@ -37,12 +37,16 @@
 #include <unordered_set>
 #include <vector>
 
-#include <zlib.h>
 
 #include "logger.h"
+#include "pdf_draw_commands.h"
+#include "pdf_font_metrics.h"
+#include "pdf_objects.h"
 #include "viewer2dcommandrenderer.h"
 
 namespace {
+using namespace layout_pdf_internal;
+
 constexpr double kLegendContentScale = 0.7;
 constexpr double kPdfPointsPerPixel = 72.0 / 96.0;
 constexpr double kLegendSymbolSize =
@@ -63,917 +67,6 @@ double ComputeTextLineAdvance(double ascent, double descent) {
   // translation. The advance mirrors the ascent + descent used by the
   // on-screen viewer when positioning multi-line labels.
   return -(ascent + descent);
-}
-
-struct TtfFontMetrics {
-  int unitsPerEm = 1000;
-  int ascent = 0;
-  int descent = 0;
-  int lineGap = 0;
-  int capHeight = 0;
-  int xMin = 0;
-  int yMin = 0;
-  int xMax = 0;
-  int yMax = 0;
-  std::array<int, 256> advanceWidths{};
-  std::array<int, 256> widths1000{};
-  std::string data;
-  bool valid = false;
-};
-
-struct PdfFontDefinition {
-  std::string key;
-  std::string family;
-  std::string baseName;
-  size_t objectId = 0;
-  bool embedded = false;
-  TtfFontMetrics metrics;
-};
-
-struct PdfFontCatalog {
-  const PdfFontDefinition *regular = nullptr;
-  const PdfFontDefinition *bold = nullptr;
-
-  const PdfFontDefinition *Resolve(const std::string &family) const;
-};
-
-std::string ToLowerCopy(std::string_view input) {
-  std::string lower;
-  lower.reserve(input.size());
-  for (char ch : input)
-    lower.push_back(static_cast<char>(std::tolower(static_cast<unsigned char>(ch))));
-  return lower;
-}
-
-const PdfFontDefinition *PdfFontCatalog::Resolve(
-    const std::string &family) const {
-  const PdfFontDefinition *fallback = regular ? regular : bold;
-  if (!fallback)
-    return nullptr;
-  if (family.empty())
-    return fallback;
-  std::string lower = ToLowerCopy(family);
-  if (bold && lower.find("bold") != std::string::npos)
-    return bold;
-  if (lower.find("sans") != std::string::npos ||
-      lower.find("arial") != std::string::npos ||
-      lower.find("dejavu") != std::string::npos)
-    return regular ? regular : bold;
-  return fallback;
-}
-
-double MeasureTextWidth(const std::string &text, double fontSize,
-                        const PdfFontDefinition *font) {
-  if (!font || !font->embedded || font->metrics.unitsPerEm <= 0)
-    return static_cast<double>(text.size()) * fontSize * 0.6;
-  double units = 0.0;
-  for (unsigned char ch : text) {
-    if (ch == '\n')
-      continue;
-    units += font->metrics.advanceWidths[ch];
-  }
-  return (units / font->metrics.unitsPerEm) * fontSize;
-}
-
-unsigned char EncodeWinAnsiCodepoint(uint32_t codepoint) {
-  if (codepoint == '\n' || codepoint == '\r' || codepoint == '\t')
-    return static_cast<unsigned char>(codepoint);
-  if (codepoint <= 0x7F)
-    return static_cast<unsigned char>(codepoint);
-  if (codepoint >= 0xA0 && codepoint <= 0xFF)
-    return static_cast<unsigned char>(codepoint);
-  switch (codepoint) {
-  case 0x20AC:
-    return 0x80;
-  case 0x201A:
-    return 0x82;
-  case 0x0192:
-    return 0x83;
-  case 0x201E:
-    return 0x84;
-  case 0x2026:
-    return 0x85;
-  case 0x2020:
-    return 0x86;
-  case 0x2021:
-    return 0x87;
-  case 0x02C6:
-    return 0x88;
-  case 0x2030:
-    return 0x89;
-  case 0x0160:
-    return 0x8A;
-  case 0x2039:
-    return 0x8B;
-  case 0x0152:
-    return 0x8C;
-  case 0x017D:
-    return 0x8E;
-  case 0x2018:
-    return 0x91;
-  case 0x2019:
-    return 0x92;
-  case 0x201C:
-    return 0x93;
-  case 0x201D:
-    return 0x94;
-  case 0x2022:
-    return 0x95;
-  case 0x2013:
-    return 0x96;
-  case 0x2014:
-    return 0x97;
-  case 0x02DC:
-    return 0x98;
-  case 0x2122:
-    return 0x99;
-  case 0x0161:
-    return 0x9A;
-  case 0x203A:
-    return 0x9B;
-  case 0x0153:
-    return 0x9C;
-  case 0x017E:
-    return 0x9E;
-  case 0x0178:
-    return 0x9F;
-  default:
-    return '?';
-  }
-}
-
-std::string EncodeWinAnsi(const std::string &utf8) {
-  std::string out;
-  out.reserve(utf8.size());
-  size_t i = 0;
-  while (i < utf8.size()) {
-    unsigned char lead = static_cast<unsigned char>(utf8[i]);
-    uint32_t codepoint = 0;
-    size_t length = 0;
-    if (lead < 0x80) {
-      codepoint = lead;
-      length = 1;
-    } else if ((lead >> 5) == 0x6 && i + 1 < utf8.size()) {
-      codepoint = ((lead & 0x1F) << 6) |
-                  (static_cast<unsigned char>(utf8[i + 1]) & 0x3F);
-      length = 2;
-    } else if ((lead >> 4) == 0xE && i + 2 < utf8.size()) {
-      codepoint = ((lead & 0x0F) << 12) |
-                  ((static_cast<unsigned char>(utf8[i + 1]) & 0x3F) << 6) |
-                  (static_cast<unsigned char>(utf8[i + 2]) & 0x3F);
-      length = 3;
-    } else if ((lead >> 3) == 0x1E && i + 3 < utf8.size()) {
-      codepoint = ((lead & 0x07) << 18) |
-                  ((static_cast<unsigned char>(utf8[i + 1]) & 0x3F) << 12) |
-                  ((static_cast<unsigned char>(utf8[i + 2]) & 0x3F) << 6) |
-                  (static_cast<unsigned char>(utf8[i + 3]) & 0x3F);
-      length = 4;
-    } else {
-      out.push_back('?');
-      ++i;
-      continue;
-    }
-    out.push_back(static_cast<char>(EncodeWinAnsiCodepoint(codepoint)));
-    i += length;
-  }
-  return out;
-}
-
-bool ReadFileToString(const std::filesystem::path &path, std::string &out) {
-  std::ifstream file(path, std::ios::binary);
-  if (!file.is_open())
-    return false;
-  std::ostringstream buffer;
-  buffer << file.rdbuf();
-  out = buffer.str();
-  return true;
-}
-
-uint16_t ReadU16(const std::string &data, size_t offset) {
-  return static_cast<uint16_t>(
-      (static_cast<unsigned char>(data[offset]) << 8) |
-      static_cast<unsigned char>(data[offset + 1]));
-}
-
-int16_t ReadS16(const std::string &data, size_t offset) {
-  return static_cast<int16_t>(ReadU16(data, offset));
-}
-
-uint32_t ReadU32(const std::string &data, size_t offset) {
-  return (static_cast<uint32_t>(static_cast<unsigned char>(data[offset])) << 24) |
-         (static_cast<uint32_t>(static_cast<unsigned char>(data[offset + 1])) << 16) |
-         (static_cast<uint32_t>(static_cast<unsigned char>(data[offset + 2])) << 8) |
-         static_cast<uint32_t>(static_cast<unsigned char>(data[offset + 3]));
-}
-
-uint32_t MakeTag(char a, char b, char c, char d) {
-  return (static_cast<uint32_t>(a) << 24) |
-         (static_cast<uint32_t>(b) << 16) |
-         (static_cast<uint32_t>(c) << 8) |
-         static_cast<uint32_t>(d);
-}
-
-bool FindTable(const std::string &data, uint32_t tag, uint32_t &offset,
-               uint32_t &length) {
-  if (data.size() < 12)
-    return false;
-  uint16_t numTables = ReadU16(data, 4);
-  size_t tableDir = 12;
-  for (uint16_t i = 0; i < numTables; ++i) {
-    size_t recordOffset = tableDir + i * 16;
-    if (recordOffset + 16 > data.size())
-      return false;
-    uint32_t entryTag = ReadU32(data, recordOffset);
-    uint32_t entryOffset = ReadU32(data, recordOffset + 8);
-    uint32_t entryLength = ReadU32(data, recordOffset + 12);
-    if (entryTag == tag) {
-      offset = entryOffset;
-      length = entryLength;
-      return entryOffset + entryLength <= data.size();
-    }
-  }
-  return false;
-}
-
-bool LoadTtfFontMetrics(const std::filesystem::path &path,
-                        TtfFontMetrics &metrics) {
-  metrics = TtfFontMetrics{};
-  std::string data;
-  if (!ReadFileToString(path, data))
-    return false;
-  if (data.size() < 12)
-    return false;
-
-  uint32_t headOffset = 0, headLength = 0;
-  uint32_t hheaOffset = 0, hheaLength = 0;
-  uint32_t maxpOffset = 0, maxpLength = 0;
-  uint32_t hmtxOffset = 0, hmtxLength = 0;
-  uint32_t cmapOffset = 0, cmapLength = 0;
-  uint32_t os2Offset = 0, os2Length = 0;
-
-  if (!FindTable(data, MakeTag('h', 'e', 'a', 'd'), headOffset, headLength))
-    return false;
-  if (!FindTable(data, MakeTag('h', 'h', 'e', 'a'), hheaOffset, hheaLength))
-    return false;
-  if (!FindTable(data, MakeTag('m', 'a', 'x', 'p'), maxpOffset, maxpLength))
-    return false;
-  if (!FindTable(data, MakeTag('h', 'm', 't', 'x'), hmtxOffset, hmtxLength))
-    return false;
-  if (!FindTable(data, MakeTag('c', 'm', 'a', 'p'), cmapOffset, cmapLength))
-    return false;
-  FindTable(data, MakeTag('O', 'S', '/', '2'), os2Offset, os2Length);
-
-  if (headOffset + 54 > data.size())
-    return false;
-  metrics.unitsPerEm = ReadU16(data, headOffset + 18);
-  metrics.xMin = ReadS16(data, headOffset + 36);
-  metrics.yMin = ReadS16(data, headOffset + 38);
-  metrics.xMax = ReadS16(data, headOffset + 40);
-  metrics.yMax = ReadS16(data, headOffset + 42);
-
-  if (hheaOffset + 36 > data.size())
-    return false;
-  metrics.ascent = ReadS16(data, hheaOffset + 4);
-  metrics.descent = ReadS16(data, hheaOffset + 6);
-  metrics.lineGap = ReadS16(data, hheaOffset + 8);
-  uint16_t numHMetrics = ReadU16(data, hheaOffset + 34);
-
-  if (maxpOffset + 6 > data.size())
-    return false;
-  uint16_t numGlyphs = ReadU16(data, maxpOffset + 4);
-  if (numGlyphs == 0)
-    return false;
-
-  if (numHMetrics == 0)
-    return false;
-
-  if (hmtxOffset + static_cast<uint32_t>(numHMetrics) * 4 > data.size())
-    return false;
-
-  std::vector<int> advanceWidths;
-  advanceWidths.resize(numGlyphs, 0);
-  int lastAdvance = 0;
-  for (uint16_t i = 0; i < numHMetrics; ++i) {
-    size_t entry = hmtxOffset + static_cast<size_t>(i) * 4;
-    int advance = ReadU16(data, entry);
-    advanceWidths[i] = advance;
-    lastAdvance = advance;
-  }
-  for (uint16_t i = numHMetrics; i < numGlyphs; ++i) {
-    advanceWidths[i] = lastAdvance;
-  }
-
-  if (os2Offset != 0 && os2Length >= 90 && os2Offset + 90 <= data.size()) {
-    uint16_t version = ReadU16(data, os2Offset);
-    if (version >= 2) {
-      metrics.capHeight = ReadS16(data, os2Offset + 88);
-    }
-  }
-  if (metrics.capHeight == 0)
-    metrics.capHeight = metrics.ascent;
-
-  const std::string cmapData =
-      data.substr(cmapOffset, std::min<size_t>(cmapLength, data.size() - cmapOffset));
-  if (cmapData.size() < 4)
-    return false;
-  uint16_t cmapTables = ReadU16(cmapData, 2);
-  size_t cmapRecordOffset = 4;
-  uint32_t chosenOffset = 0;
-  for (uint16_t i = 0; i < cmapTables; ++i) {
-    if (cmapRecordOffset + 8 > cmapData.size())
-      return false;
-    uint16_t platformId = ReadU16(cmapData, cmapRecordOffset);
-    uint16_t encodingId = ReadU16(cmapData, cmapRecordOffset + 2);
-    uint32_t subOffset = ReadU32(cmapData, cmapRecordOffset + 4);
-    size_t subBase = subOffset;
-    if (subBase + 2 > cmapData.size())
-      continue;
-    uint16_t format = ReadU16(cmapData, subBase);
-    if (format == 4 && platformId == 3 &&
-        (encodingId == 1 || encodingId == 0)) {
-      chosenOffset = subOffset;
-      break;
-    }
-    cmapRecordOffset += 8;
-  }
-  if (chosenOffset == 0)
-    return false;
-
-  size_t subBase = chosenOffset;
-  if (subBase + 14 > cmapData.size())
-    return false;
-  uint16_t segCount = ReadU16(cmapData, subBase + 6) / 2;
-  size_t endCountOffset = subBase + 14;
-  size_t startCountOffset = endCountOffset + 2 * segCount + 2;
-  size_t idDeltaOffset = startCountOffset + 2 * segCount;
-  size_t idRangeOffsetOffset = idDeltaOffset + 2 * segCount;
-  if (idRangeOffsetOffset + 2 * segCount > cmapData.size())
-    return false;
-
-  auto glyphForCodepoint = [&](uint16_t code) -> uint16_t {
-    for (uint16_t i = 0; i < segCount; ++i) {
-      uint16_t endCount = ReadU16(cmapData, endCountOffset + 2 * i);
-      uint16_t startCount = ReadU16(cmapData, startCountOffset + 2 * i);
-      if (code < startCount || code > endCount)
-        continue;
-      int16_t idDelta = ReadS16(cmapData, idDeltaOffset + 2 * i);
-      uint16_t idRangeOffset = ReadU16(cmapData, idRangeOffsetOffset + 2 * i);
-      if (idRangeOffset == 0) {
-        return static_cast<uint16_t>(code + idDelta);
-      }
-      size_t glyphOffset =
-          idRangeOffsetOffset + 2 * i + idRangeOffset + 2 * (code - startCount);
-      if (glyphOffset + 2 > cmapData.size())
-        return 0;
-      uint16_t glyphIndex = ReadU16(cmapData, glyphOffset);
-      if (glyphIndex == 0)
-        return 0;
-      return static_cast<uint16_t>(glyphIndex + idDelta);
-    }
-    return 0;
-  };
-
-  int missingWidth = advanceWidths.empty() ? 0 : advanceWidths[0];
-  for (size_t i = 0; i < metrics.advanceWidths.size(); ++i) {
-    uint16_t glyphIndex = glyphForCodepoint(static_cast<uint16_t>(i));
-    int advance = missingWidth;
-    if (glyphIndex < advanceWidths.size())
-      advance = advanceWidths[glyphIndex];
-    metrics.advanceWidths[i] = advance;
-    if (metrics.unitsPerEm > 0) {
-      metrics.widths1000[i] =
-          static_cast<int>(std::lround(advance * 1000.0 / metrics.unitsPerEm));
-    } else {
-      metrics.widths1000[i] = 0;
-    }
-  }
-
-  metrics.data = std::move(data);
-  metrics.valid = metrics.unitsPerEm > 0;
-  return metrics.valid;
-}
-
-std::filesystem::path FindFontPath(bool bold) {
-  struct FontCandidate {
-    const char *regularPath;
-    const char *boldPath;
-  };
-  // Keep this list in sync with the font face names used by the UI legend/event
-  // table rendering so PDFs and on-screen views share the same family.
-  const std::vector<FontCandidate> candidates = {
-#ifdef _WIN32
-      {"C:/Windows/Fonts/arial.ttf", "C:/Windows/Fonts/arialbd.ttf"},
-#elif defined(__APPLE__)
-      {"/Library/Fonts/Arial.ttf", "/Library/Fonts/Arial Bold.ttf"},
-      {"/System/Library/Fonts/Supplemental/Arial.ttf",
-       "/System/Library/Fonts/Supplemental/Arial Bold.ttf"},
-#else
-      {"/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf",
-       "/usr/share/fonts/truetype/dejavu/DejaVuSans-Bold.ttf"},
-      {"/usr/share/fonts/truetype/liberation/LiberationSans-Regular.ttf",
-       "/usr/share/fonts/truetype/liberation/LiberationSans-Bold.ttf"},
-#endif
-  };
-  for (const auto &candidate : candidates) {
-    const char *path = bold ? candidate.boldPath : candidate.regularPath;
-    if (path && std::filesystem::exists(path))
-      return std::filesystem::path(path);
-  }
-  return {};
-}
-
-bool LoadPdfFontMetrics(PdfFontDefinition &font, bool bold) {
-  std::filesystem::path path = FindFontPath(bold);
-  if (path.empty())
-    return false;
-  return LoadTtfFontMetrics(path, font.metrics);
-}
-
-struct PdfObject {
-  std::string body;
-};
-
-bool AppendEmbeddedFontObjects(std::vector<PdfObject> &objects,
-                               PdfFontDefinition &font) {
-  if (!font.metrics.valid || font.metrics.data.empty())
-    return false;
-  const double scale = font.metrics.unitsPerEm > 0
-                           ? 1000.0 / font.metrics.unitsPerEm
-                           : 1.0;
-  int ascent = static_cast<int>(std::lround(font.metrics.ascent * scale));
-  int descent =
-      -static_cast<int>(std::lround(std::abs(font.metrics.descent) * scale));
-  int capHeight = static_cast<int>(std::lround(font.metrics.capHeight * scale));
-  int xMin = static_cast<int>(std::lround(font.metrics.xMin * scale));
-  int yMin = static_cast<int>(std::lround(font.metrics.yMin * scale));
-  int xMax = static_cast<int>(std::lround(font.metrics.xMax * scale));
-  int yMax = static_cast<int>(std::lround(font.metrics.yMax * scale));
-
-  size_t fontFileIndex = objects.size() + 1;
-  std::ostringstream fontFileStream;
-  const bool needsNewline =
-      font.metrics.data.empty() || font.metrics.data.back() != '\n';
-  const size_t streamLength =
-      font.metrics.data.size() + (needsNewline ? 1u : 0u);
-  fontFileStream << "<< /Length " << streamLength << " /Length1 "
-                 << font.metrics.data.size() << " >>\nstream\n"
-                 << font.metrics.data;
-  if (needsNewline)
-    fontFileStream << '\n';
-  fontFileStream << "endstream";
-  objects.push_back({fontFileStream.str()});
-
-  size_t descriptorIndex = objects.size() + 1;
-  std::ostringstream descriptor;
-  descriptor << "<< /Type /FontDescriptor /FontName /" << font.baseName
-             << " /Flags 32 /FontBBox [" << xMin << ' ' << yMin << ' '
-             << xMax << ' ' << yMax << "] /Ascent " << ascent
-             << " /Descent " << descent << " /CapHeight " << capHeight
-             << " /ItalicAngle 0 /StemV 80 /FontFile2 " << fontFileIndex
-             << " 0 R >>";
-  objects.push_back({descriptor.str()});
-
-  size_t fontIndex = objects.size() + 1;
-  std::ostringstream fontObject;
-  fontObject << "<< /Type /Font /Subtype /TrueType /BaseFont /"
-             << font.baseName << " /FirstChar 32 /LastChar 255 /Widths [";
-  for (int code = 32; code <= 255; ++code) {
-    fontObject << font.metrics.widths1000[static_cast<unsigned char>(code)];
-    if (code != 255)
-      fontObject << ' ';
-  }
-  fontObject << "] /FontDescriptor " << descriptorIndex
-             << " 0 R /Encoding /WinAnsiEncoding >>";
-  objects.push_back({fontObject.str()});
-
-  font.objectId = fontIndex;
-  font.embedded = true;
-  return true;
-}
-
-void AppendFallbackType1Font(std::vector<PdfObject> &objects,
-                             PdfFontDefinition &font,
-                             const std::string &baseFont) {
-  objects.push_back(
-      {"<< /Type /Font /Subtype /Type1 /BaseFont /" + baseFont + " >>"});
-  font.objectId = objects.size();
-  font.embedded = false;
-  font.baseName = baseFont;
-}
-
-class FloatFormatter {
-public:
-  explicit FloatFormatter(int precision)
-      : precision_(std::clamp(precision, 0, 6)) {}
-
-  std::string Format(double value) const {
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(precision_) << value;
-    return ss.str();
-  }
-
-private:
-  int precision_;
-};
-
-class PdfDeflater {
-public:
-  static bool Compress(const std::string &input, std::string &output,
-                       std::string &error) {
-    if (input.empty()) {
-      output.clear();
-      return true;
-    }
-
-    uLongf bound = compressBound(input.size());
-    std::string compressed;
-    compressed.resize(bound);
-
-    int zres = compress2(reinterpret_cast<Bytef *>(compressed.data()), &bound,
-                         reinterpret_cast<const Bytef *>(input.data()),
-                         input.size(), Z_BEST_SPEED);
-    if (zres != Z_OK) {
-      error = "compress2 failed";
-      return false;
-    }
-
-    compressed.resize(bound);
-    output.swap(compressed);
-    return true;
-  }
-};
-
-struct Point {
-  double x = 0.0;
-  double y = 0.0;
-};
-
-struct Transform {
-  double scale = 1.0;
-  double offsetX = 0.0;
-  double offsetY = 0.0;
-};
-
-struct Mapping {
-  double minX = 0.0;
-  double minY = 0.0;
-  double scale = 1.0;
-  double offsetX = 0.0;
-  double offsetY = 0.0;
-  double drawHeight = 0.0;
-  bool flipY = true;
-};
-
-struct RenderOptions {
-  bool includeText = true;
-  const std::unordered_map<std::string, std::string> *symbolKeyNames = nullptr;
-  const std::unordered_map<uint32_t, std::string> *symbolIdNames = nullptr;
-  const PdfFontCatalog *fonts = nullptr;
-  double strokeScale = 1.0;
-};
-
-Point Apply(const Transform &t, double x, double y) {
-  return {x * t.scale + t.offsetX, y * t.scale + t.offsetY};
-}
-
-Point MapWithMapping(double x, double y, const Mapping &mapping) {
-  double px = mapping.offsetX + (x - mapping.minX) * mapping.scale;
-  double py = mapping.offsetY + (y - mapping.minY) * mapping.scale;
-  if (mapping.flipY)
-    py = mapping.offsetY + mapping.drawHeight -
-         (y - mapping.minY) * mapping.scale;
-  return {px, py};
-}
-
-class GraphicsStateCache {
-public:
-  void SetStroke(std::ostringstream &out, const CanvasStroke &stroke,
-                 const FloatFormatter &fmt) {
-    if (!joinStyleSet_) {
-      out << "1 j\n";
-      joinStyleSet_ = true;
-    }
-    if (!capStyleSet_) {
-      out << "1 J\n";
-      capStyleSet_ = true;
-    }
-    if (!hasStrokeColor_ || !SameColor(stroke.color, strokeColor_)) {
-      out << fmt.Format(stroke.color.r) << ' ' << fmt.Format(stroke.color.g) << ' '
-          << fmt.Format(stroke.color.b) << " RG\n";
-      strokeColor_ = stroke.color;
-      hasStrokeColor_ = true;
-    }
-    if (!hasLineWidth_ || std::abs(stroke.width - lineWidth_) > 1e-6) {
-      out << fmt.Format(stroke.width) << " w\n";
-      lineWidth_ = stroke.width;
-      hasLineWidth_ = true;
-    }
-  }
-
-  void SetFill(std::ostringstream &out, const CanvasFill &fill,
-               const FloatFormatter &fmt) {
-    if (!hasFillColor_ || !SameColor(fill.color, fillColor_)) {
-      out << fmt.Format(fill.color.r) << ' ' << fmt.Format(fill.color.g) << ' '
-          << fmt.Format(fill.color.b) << " rg\n";
-      fillColor_ = fill.color;
-      hasFillColor_ = true;
-    }
-  }
-
-private:
-  static bool SameColor(const CanvasColor &a, const CanvasColor &b) {
-    return std::abs(a.r - b.r) < 1e-6 && std::abs(a.g - b.g) < 1e-6 &&
-           std::abs(a.b - b.b) < 1e-6;
-  }
-
-  CanvasColor strokeColor_{};
-  CanvasColor fillColor_{};
-  double lineWidth_ = -1.0;
-  bool hasStrokeColor_ = false;
-  bool hasFillColor_ = false;
-  bool hasLineWidth_ = false;
-  bool joinStyleSet_ = false;
-  bool capStyleSet_ = false;
-};
-
-void AppendLine(std::ostringstream &out, GraphicsStateCache &cache,
-                const FloatFormatter &fmt, const Point &a, const Point &b,
-                const CanvasStroke &stroke) {
-  cache.SetStroke(out, stroke, fmt);
-  out << fmt.Format(a.x) << ' ' << fmt.Format(a.y) << " m\n"
-      << fmt.Format(b.x) << ' ' << fmt.Format(b.y) << " l\nS\n";
-}
-
-void AppendPolyline(std::ostringstream &out, GraphicsStateCache &cache,
-                    const FloatFormatter &fmt, const std::vector<Point> &pts,
-                    const CanvasStroke &stroke) {
-  if (pts.size() < 2)
-    return;
-  cache.SetStroke(out, stroke, fmt);
-  out << fmt.Format(pts[0].x) << ' ' << fmt.Format(pts[0].y) << " m\n";
-  for (size_t i = 1; i < pts.size(); ++i) {
-    out << fmt.Format(pts[i].x) << ' ' << fmt.Format(pts[i].y) << " l\n";
-  }
-  out << "S\n";
-}
-
-void AppendPolygon(std::ostringstream &out, GraphicsStateCache &cache,
-                   const FloatFormatter &fmt, const std::vector<Point> &pts,
-                   const CanvasStroke &stroke, const CanvasFill *fill) {
-  if (pts.size() < 3)
-    return;
-  auto emitPath = [&]() {
-    out << fmt.Format(pts[0].x) << ' ' << fmt.Format(pts[0].y) << " m\n";
-    for (size_t i = 1; i < pts.size(); ++i)
-      out << fmt.Format(pts[i].x) << ' ' << fmt.Format(pts[i].y) << " l\n";
-    out << "h\n";
-  };
-
-  if (stroke.width > 0.0f) {
-    cache.SetStroke(out, stroke, fmt);
-    emitPath();
-    out << "S\n";
-  }
-
-  if (fill) {
-    cache.SetFill(out, *fill, fmt);
-    emitPath();
-    out << "f\n";
-  }
-}
-
-void AppendRectangle(std::ostringstream &out, GraphicsStateCache &cache,
-                     const FloatFormatter &fmt, const Point &origin, double w,
-                     double h, const CanvasStroke &stroke,
-                     const CanvasFill *fill) {
-  auto emitRect = [&]() {
-    out << fmt.Format(origin.x) << ' ' << fmt.Format(origin.y) << ' '
-        << fmt.Format(w) << ' ' << fmt.Format(h) << " re\n";
-  };
-
-  if (stroke.width > 0.0f) {
-    cache.SetStroke(out, stroke, fmt);
-    emitRect();
-    out << "S\n";
-  }
-
-  if (fill) {
-    cache.SetFill(out, *fill, fmt);
-    emitRect();
-    out << "f\n";
-  }
-}
-
-void AppendCircle(std::ostringstream &out, GraphicsStateCache &cache,
-                  const FloatFormatter &fmt, const Point &center,
-                  double radius, const CanvasStroke &stroke,
-                  const CanvasFill *fill) {
-  // Approximate circle with 4 cubic Beziers.
-  const double c = radius * 0.552284749831; // 4*(sqrt(2)-1)/3
-  Point p0{center.x + radius, center.y};
-  Point p1{center.x + radius, center.y + c};
-  Point p2{center.x + c, center.y + radius};
-  Point p3{center.x, center.y + radius};
-  Point p4{center.x - c, center.y + radius};
-  Point p5{center.x - radius, center.y + c};
-  Point p6{center.x - radius, center.y};
-  Point p7{center.x - radius, center.y - c};
-  Point p8{center.x - c, center.y - radius};
-  Point p9{center.x, center.y - radius};
-  Point p10{center.x + c, center.y - radius};
-  Point p11{center.x + radius, center.y - c};
-
-  auto emitCircle = [&]() {
-    out << fmt.Format(p0.x) << ' ' << fmt.Format(p0.y) << " m\n"
-        << fmt.Format(p1.x) << ' ' << fmt.Format(p1.y) << ' '
-        << fmt.Format(p2.x) << ' ' << fmt.Format(p2.y) << ' '
-        << fmt.Format(p3.x) << ' ' << fmt.Format(p3.y) << " c\n"
-        << fmt.Format(p4.x) << ' ' << fmt.Format(p4.y) << ' '
-        << fmt.Format(p5.x) << ' ' << fmt.Format(p5.y) << ' '
-        << fmt.Format(p6.x) << ' ' << fmt.Format(p6.y) << " c\n"
-        << fmt.Format(p7.x) << ' ' << fmt.Format(p7.y) << ' '
-        << fmt.Format(p8.x) << ' ' << fmt.Format(p8.y) << ' '
-        << fmt.Format(p9.x) << ' ' << fmt.Format(p9.y) << " c\n"
-        << fmt.Format(p10.x) << ' ' << fmt.Format(p10.y) << ' '
-        << fmt.Format(p11.x) << ' ' << fmt.Format(p11.y) << ' '
-        << fmt.Format(p0.x) << ' ' << fmt.Format(p0.y) << " c\n";
-  };
-
-  if (stroke.width > 0.0f) {
-    cache.SetStroke(out, stroke, fmt);
-    emitCircle();
-    out << "S\n";
-  }
-
-  if (fill) {
-    cache.SetFill(out, *fill, fmt);
-    emitCircle();
-    out << "f\n";
-  }
-}
-
-void AppendText(std::ostringstream &out, const FloatFormatter &fmt,
-                const Point &pos, const TextCommand &cmd,
-                const CanvasTextStyle &style, double scale,
-                const PdfFontCatalog *fonts) {
-  const std::string encodedText = EncodeWinAnsi(cmd.text);
-  const PdfFontDefinition *font =
-      fonts ? fonts->Resolve(style.fontFamily) : nullptr;
-  double scaledFontSize = style.fontSize * scale;
-  if (font && font->embedded && font->metrics.unitsPerEm > 0 &&
-      style.ascent > 0.0f && style.descent > 0.0f) {
-    const double targetHeight = (style.ascent + style.descent) * scale;
-    const double fontHeightUnits =
-        font->metrics.ascent + std::abs(font->metrics.descent);
-    if (fontHeightUnits > 0.0) {
-      const double fontHeight =
-          fontHeightUnits * scaledFontSize / font->metrics.unitsPerEm;
-      if (fontHeight > 0.0)
-        scaledFontSize *= targetHeight / fontHeight;
-    }
-  }
-
-  auto measureLineWidth = [&](std::string_view line) {
-    if (!font || !font->embedded)
-      return static_cast<double>(line.size()) * scaledFontSize * 0.6;
-    double units = 0.0;
-    for (unsigned char ch : line) {
-      units += font->metrics.advanceWidths[ch];
-    }
-    return (units / font->metrics.unitsPerEm) * scaledFontSize;
-  };
-  const double fallbackAscent =
-      font && font->embedded
-          ? (font->metrics.ascent * scaledFontSize / font->metrics.unitsPerEm)
-          : scaledFontSize * 0.8;
-  const double fallbackDescent =
-      font && font->embedded
-          ? (std::abs(font->metrics.descent) * scaledFontSize /
-             font->metrics.unitsPerEm)
-          : scaledFontSize * 0.2;
-  const double ascent =
-      style.ascent > 0.0f ? style.ascent * scale : fallbackAscent;
-  const double descent =
-      style.descent > 0.0f ? style.descent * scale : fallbackDescent;
-  const double measuredLineHeight =
-      style.lineHeight > 0.0f
-          ? style.lineHeight * scale
-          : (ascent + descent +
-             (font && font->embedded
-                  ? (font->metrics.lineGap * scaledFontSize /
-                     font->metrics.unitsPerEm)
-                  : 0.0));
-  const double extraSpacing =
-      style.lineHeight > 0.0f ? style.extraLineSpacing * scale : 0.0;
-
-  double maxLineWidth = 0.0;
-  size_t lineStart = 0;
-  for (size_t i = 0; i <= encodedText.size(); ++i) {
-    if (i == encodedText.size() || encodedText[i] == '\n') {
-      maxLineWidth = std::max(
-          maxLineWidth,
-          measureLineWidth(std::string_view(encodedText)
-                               .substr(lineStart, i - lineStart)));
-      lineStart = i + 1;
-    }
-  }
-
-  double horizontalOffset = 0.0;
-  if (style.hAlign == CanvasTextStyle::HorizontalAlign::Center)
-    horizontalOffset = -maxLineWidth / 2.0;
-  else if (style.hAlign == CanvasTextStyle::HorizontalAlign::Right)
-    horizontalOffset = -maxLineWidth;
-
-  double verticalOffset = 0.0;
-  switch (style.vAlign) {
-  case CanvasTextStyle::VerticalAlign::Top:
-    verticalOffset = -ascent;
-    break;
-  case CanvasTextStyle::VerticalAlign::Middle:
-    verticalOffset = -(ascent - descent) * 0.5;
-    break;
-  case CanvasTextStyle::VerticalAlign::Bottom:
-    verticalOffset = descent;
-    break;
-  case CanvasTextStyle::VerticalAlign::Baseline:
-    break;
-  }
-
-  // Always advance downward for successive lines to mirror the on-screen
-  // rendering, even if upstream metrics change sign conventions.
-  double lineAdvance = 0.0;
-  if (style.lineHeight > 0.0f) {
-    lineAdvance = -(measuredLineHeight + extraSpacing);
-  } else {
-    lineAdvance = ComputeTextLineAdvance(ascent, descent);
-  }
-  if (lineAdvance > 0.0)
-    lineAdvance = -lineAdvance;
-  auto emitText = [&](const CanvasColor &color, double dx, double dy) {
-    const char *fontKey = font ? font->key.c_str() : "F1";
-    out << "BT\n/" << fontKey << ' ' << fmt.Format(scaledFontSize) << " Tf\n";
-    out << fmt.Format(color.r) << ' ' << fmt.Format(color.g) << ' '
-        << fmt.Format(color.b) << " rg\n";
-    out << fmt.Format(pos.x + horizontalOffset + dx) << ' '
-        << fmt.Format(pos.y + verticalOffset + dy) << " Td\n";
-    out << "(";
-    for (char ch : encodedText) {
-      if (ch == '(' || ch == ')' || ch == '\\')
-        out << '\\';
-      if (ch == '\n') {
-        out << ") Tj\n0 " << fmt.Format(lineAdvance) << " Td\n(";
-        continue;
-      }
-      out << ch;
-    }
-    out << ") Tj\nET\n";
-  };
-
-  const double outline = style.outlineWidth * scale;
-  if (outline > 0.0f) {
-    const std::array<std::array<double, 2>, 8> offsets = {
-        std::array<double, 2>{-outline, 0.0},
-        std::array<double, 2>{outline, 0.0},
-        std::array<double, 2>{0.0, -outline},
-        std::array<double, 2>{0.0, outline},
-        std::array<double, 2>{-outline, -outline},
-        std::array<double, 2>{outline, -outline},
-        std::array<double, 2>{-outline, outline},
-        std::array<double, 2>{outline, outline}};
-    for (const auto &offset : offsets)
-      emitText(style.outlineColor, offset[0], offset[1]);
-  }
-
-  emitText(style.color, 0.0, 0.0);
-}
-
-Point MapPointWithTransform(double x, double y, const Transform &current,
-                            const Mapping &mapping) {
-  auto applied = Apply(current, x, y);
-  return MapWithMapping(applied.x, applied.y, mapping);
-}
-
-Transform2D TransformFromCanvas(const CanvasTransform &transform) {
-  Transform2D out{};
-  out.a = transform.scale;
-  out.d = transform.scale;
-  out.tx = transform.offsetX;
-  out.ty = transform.offsetY;
-  return out;
-}
-
-void AppendSymbolInstance(std::ostringstream &out, const FloatFormatter &fmt,
-                          const Mapping &mapping,
-                          const Transform2D &transform,
-                          const std::string &name) {
-  double translateX = mapping.scale * transform.tx +
-                      mapping.offsetX - mapping.minX * mapping.scale;
-  double translateY = mapping.scale * transform.ty +
-                      mapping.offsetY - mapping.minY * mapping.scale;
-  out << "q\n" << fmt.Format(transform.a) << ' ' << fmt.Format(transform.b)
-      << ' ' << fmt.Format(transform.c) << ' ' << fmt.Format(transform.d)
-      << ' ' << fmt.Format(translateX) << ' ' << fmt.Format(translateY)
-      << " cm\n/" << name << " Do\nQ\n";
 }
 
 int SymbolViewRank(SymbolViewKind kind) {
@@ -1107,100 +200,6 @@ SymbolBounds ComputeSymbolBounds(const std::vector<CanvasCommand> &commands) {
 // fills in separate functions allows the caller to control layering
 // explicitly, which is required to match the on-screen 2D viewer where fills
 // occlude internal wireframe edges within the same group.
-void EmitCommandStroke(std::ostringstream &content, GraphicsStateCache &cache,
-                       const FloatFormatter &formatter, const Mapping &mapping,
-                       const Transform &current, const CanvasCommand &command,
-                       const RenderOptions &options) {
-  const double strokeScale = mapping.scale * options.strokeScale;
-  std::visit(
-      [&](auto &&c) {
-        using T = std::decay_t<decltype(c)>;
-        if constexpr (std::is_same_v<T, LineCommand>) {
-          auto pa = MapPointWithTransform(c.x0, c.y0, current, mapping);
-          auto pb = MapPointWithTransform(c.x1, c.y1, current, mapping);
-          CanvasStroke stroke = c.stroke;
-          stroke.width *= strokeScale;
-          AppendLine(content, cache, formatter, pa, pb, stroke);
-        } else if constexpr (std::is_same_v<T, PolylineCommand>) {
-          std::vector<Point> pts;
-          pts.reserve(c.points.size() / 2);
-          for (size_t i = 0; i + 1 < c.points.size(); i += 2)
-            pts.push_back(
-                MapPointWithTransform(c.points[i], c.points[i + 1], current,
-                                      mapping));
-          CanvasStroke stroke = c.stroke;
-          stroke.width *= strokeScale;
-          AppendPolyline(content, cache, formatter, pts, stroke);
-        } else if constexpr (std::is_same_v<T, PolygonCommand>) {
-          std::vector<Point> pts;
-          pts.reserve(c.points.size() / 2);
-          for (size_t i = 0; i + 1 < c.points.size(); i += 2)
-            pts.push_back(
-                MapPointWithTransform(c.points[i], c.points[i + 1], current,
-                                      mapping));
-          CanvasStroke stroke = c.stroke;
-          stroke.width *= strokeScale;
-          AppendPolygon(content, cache, formatter, pts, stroke, nullptr);
-        } else if constexpr (std::is_same_v<T, RectangleCommand>) {
-          auto origin = MapPointWithTransform(c.x, c.y, current, mapping);
-          double w = c.w * current.scale * mapping.scale;
-          double h = c.h * current.scale * mapping.scale;
-          CanvasStroke stroke = c.stroke;
-          stroke.width *= strokeScale;
-          AppendRectangle(content, cache, formatter, origin, w, h, stroke,
-                          nullptr);
-        } else if constexpr (std::is_same_v<T, CircleCommand>) {
-          auto center = MapPointWithTransform(c.cx, c.cy, current, mapping);
-          double radius = c.radius * current.scale * mapping.scale;
-          CanvasStroke stroke = c.stroke;
-          stroke.width *= strokeScale;
-          AppendCircle(content, cache, formatter, center, radius, stroke,
-                       nullptr);
-        }
-      },
-      command);
-}
-
-// Emits only the fill portion of a drawing command. Stroke width is forced to
-// zero to ensure no outlines leak back in when rendering fills as a separate
-// pass.
-void EmitCommandFill(std::ostringstream &content, GraphicsStateCache &cache,
-                     const FloatFormatter &formatter, const Mapping &mapping,
-                     const Transform &current, const CanvasCommand &command) {
-  std::visit(
-      [&](auto &&c) {
-        using T = std::decay_t<decltype(c)>;
-        if constexpr (std::is_same_v<T, PolygonCommand>) {
-          std::vector<Point> pts;
-          pts.reserve(c.points.size() / 2);
-          for (size_t i = 0; i + 1 < c.points.size(); i += 2)
-            pts.push_back(
-                MapPointWithTransform(c.points[i], c.points[i + 1], current,
-                                      mapping));
-          CanvasStroke disabledStroke = c.stroke;
-          disabledStroke.width = 0.0f;
-          AppendPolygon(content, cache, formatter, pts, disabledStroke,
-                        &c.fill);
-        } else if constexpr (std::is_same_v<T, RectangleCommand>) {
-          auto origin = MapPointWithTransform(c.x, c.y, current, mapping);
-          double w = c.w * current.scale * mapping.scale;
-          double h = c.h * current.scale * mapping.scale;
-          CanvasStroke disabledStroke = c.stroke;
-          disabledStroke.width = 0.0f;
-          AppendRectangle(content, cache, formatter, origin, w, h,
-                          disabledStroke, &c.fill);
-        } else if constexpr (std::is_same_v<T, CircleCommand>) {
-          auto center = MapPointWithTransform(c.cx, c.cy, current, mapping);
-          double radius = c.radius * current.scale * mapping.scale;
-          CanvasStroke disabledStroke = c.stroke;
-          disabledStroke.width = 0.0f;
-          AppendCircle(content, cache, formatter, center, radius,
-                       disabledStroke, &c.fill);
-        }
-      },
-      command);
-}
-
 std::string RenderCommandsToStream(
     const std::vector<CanvasCommand> &commands,
     const std::vector<CommandMetadata> &metadata,
